Add several more tests for _to_database

+++ src/py4vasp/_calculation/force.py
@@ -124,14 +124,14 @@ POSITION                                       TOTAL-FORCE (eV/Angst)
         initial_force_norms = np.linalg.norm(self._force[0], axis=-1)
         return {
             "force": {
-                "final_force_min": np.min(final_force_norms),
-                "final_force_median": np.median(final_force_norms),
-                "final_force_mean": np.mean(final_force_norms),
-                "final_force_max": np.max(final_force_norms),
-                "final_index_force_max": np.argmax(final_force_norms),
-                "initial_force_min": np.min(initial_force_norms),
-                "initial_force_max": np.max(initial_force_norms),
-                "initial_index_force_max": np.argmax(initial_force_norms),
+                "final_force_min": float(np.min(final_force_norms)),
+                "final_force_median": float(np.median(final_force_norms)),
+                "final_force_mean": float(np.mean(final_force_norms)),
+                "final_force_max": float(np.max(final_force_norms)),
+                "final_index_force_max": int(np.argmax(final_force_norms)),
+                "initial_force_min": float(np.min(initial_force_norms)),
+                "initial_force_max": float(np.max(initial_force_norms)),
+                "initial_index_force_max": int(np.argmax(initial_force_norms)),
             },
         }
